requirements.txt
alembic==0.9.8
aliyun-python-sdk-core-v3==2.8.6
aliyun-python-sdk-dysmsapi==1.0.0
aniso8601==3.0.0
beautifulsoup4==4.6.1
blinker==1.4
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
crcmod==1.7
dukpy==0.2.0
echarts-themes-pypkg==0.0.3
Flask==0.12.2
Flask-Caching==1.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
flask-paginate==0.5.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
html5lib==1.0.1
idna==2.6
itsdangerous==0.24
javascripthon==0.10
Jinja2==2.10
jupyter-echarts-pypkg==0.1.2
lml==0.0.2
lxml==4.2.1
macropy3==1.1.0b2
Mako==1.0.7
MarkupSafe==1.0

oss2==2.4.0
Pillow==5.1.0
pyecharts==0.5.6
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pymongo==3.7.1
PyMySQL==0.8.0
python-dateutil==2.7.0
python-editor==1.0.3
python-memcached==1.59
pytz==2018.3
qiniu==7.2.0
redis==2.10.6
requests==2.18.4
shortuuid==0.5.0
six==1.11.0
SQLAlchemy==1.2.5
urllib3==1.22
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.1
